==> 06_network_analysis.Rmd <==
# Network analysis

To run this notebook, you need to make sure you have first run:
1. [`index.Rmd`](./index.Rmd)
2. [`01_data_preparation.Rmd`](./01_data_preparation.Rmd), and 
3. from [`02_mag_catalogue.Rmd`](./02_mag_catalogue.Rmd) to have access to the `functional_distances`

Also, you need to load a set of custom functions to parse the `.cx` format networks. 
`microbetag` returns `.cx2` networks, but we could not find a more straight forward way to parse them in R, so after loading those on Cytoscape, we saved them as `.cx`.

```{r load cx parser}
source("loadcx.R")

order_colors <- c(
  "#08D1D1","#BF7022","#2EBC82","#2E91BC","#2E5ABC","#09CE8D","#E0A608","#097265","#AB56D8","#0AA01F","#274945","#4E397A","#B3265B",
  "#2EBC45","#7F4561","#CE7E00","#BC2EA1","#6A52C6","#AD8D8D", "#CFE71F")

```




## Overall 


```{r overvall}
cxFile  <- "microbetag_analysis/microbetag_nets/cx/overall.cx"
overall <- process_cxfile(cxFile, "overall", genome_metadata, genome_counts_filt, order_colors)
```


```{r plots-overall}
all_days_pos_plot <- plot_graph(overall$graph_pos, overall$clusters_pos, "microbetag_analysis/figs/flashweave_scores/overall_pos.svg")
all_days_neg_plot <- plot_graph(overall$graph_neg, overall$clusters_neg, "microbetag_analysis/figs/flashweave_scores/overall_neg.svg")
```


## Timewise analysis

### Day 7

The `threshold` refers to the minimum absolute flashweave score. By default, we use 0.5 as coming from microbetag's configuration. 

```{r cor_day7, message=FALSE, warning=FALSE, comment=""}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_day_7.cx"
day7   <- process_cxfile(cxFile, "day7", genome_metadata, genome_counts_filt, order_colors, col_name="day")
```


```{r cor_day7_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
d7_pos_plot <- plot_graph(day7$graph_pos, day7$clusters_pos, "microbetag_analysis/figs/flashweave_scores/day07_pos.svg")
d7_neg_plot <- plot_graph(day7$graph_neg, day7$clusters_neg, "microbetag_analysis/figs/flashweave_scores/day07_neg.svg")
```



### Day 14

```{r cor_day14, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_day_14.cx"
day14  <- process_cxfile(cxFile, "day14", genome_metadata, genome_counts_filt, order_colors, col_name = "day")
```

```{r cor_day14_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
d14_pos_plot <- plot_graph(day14$graph_pos, day14$clusters_pos, "microbetag_analysis/figs/flashweave_scores/day14_pos.svg")
d14_neg_plot <- plot_graph(day14$graph_neg, day14$clusters_neg, "microbetag_analysis/figs/flashweave_scores/day14_neg.svg")
```

### Day 21

```{r cor_day21, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_day_21.cx"
day21  <- process_cxfile(cxFile, "day21", genome_metadata, genome_counts_filt, order_colors, col_name = "day")
```


```{r cor_day21_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
d21_pos_plot <- plot_graph(day21$graph_pos, day21$clusters_pos, "microbetag_analysis/figs/flashweave_scores/day21_pos.svg")
d21_neg_plot <- plot_graph(day21$graph_neg, day21$clusters_neg, "microbetag_analysis/figs/flashweave_scores/day21_neg.svg")
```

### Day 28

```{r cor_day28, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_day_28.cx"
day28  <- process_cxfile(cxFile, "day28", genome_metadata, genome_counts_filt, order_colors, col_name = "day")
```


```{r cor_day28_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
d28_pos_plot <- plot_graph(day28$graph_pos, day28$clusters_pos, "microbetag_analysis/figs/flashweave_scores/day28_pos.svg")
d28_neg_plot <- plot_graph(day28$graph_neg, day28$clusters_neg, "microbetag_analysis/figs/flashweave_scores/day28_neg.svg")
```

### Day 35

```{r cor_day35, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_day_35.cx"
day35  <- process_cxfile(cxFile, "day35", genome_metadata, genome_counts_filt, order_colors, col_name = "day")
```


```{r cor_day35_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
d35_pos_plot <- plot_graph(day35$graph_pos, day35$clusters_pos, "microbetag_analysis/figs/flashweave_scores/day35_pos.svg")
d35_neg_plot <- plot_graph(day35$graph_neg, day35$clusters_neg, "microbetag_analysis/figs/flashweave_scores/day35_neg.svg")
```



## Treatment-wise analysis

### TG1

```{r cor_TG1, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_TG1.cx"
TG1  <- process_cxfile(cxFile, "TG1", genome_metadata, genome_counts_filt, order_colors, col_name = "treatment")
```


```{r cor_TG1_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
TG1_pos_plot <- plot_graph(TG1$graph_pos, TG1$clusters_pos, "microbetag_analysis/figs/flashweave_scores/TG1_pos.svg")
TG1_neg_plot <- plot_graph(TG1$graph_neg, TG1$clusters_neg, "microbetag_analysis/figs/flashweave_scores/TG1_neg.svg")
```

### TG2

```{r cor_TG2, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_TG2.cx"
TG2    <- process_cxfile(cxFile, "TG2", genome_metadata, genome_counts_filt, order_colors, col_name = "treatment")
```

```{r cor_TG2_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
TG2_pos_plot <- plot_graph(TG2$graph_pos, TG2$clusters_pos, "microbetag_analysis/figs/flashweave_scores/TG2_pos.svg")
TG2_neg_plot <- plot_graph(TG2$graph_neg, TG2$clusters_neg, "microbetag_analysis/figs/flashweave_scores/TG2_neg.svg")
```

### TG3

```{r cor_TG3, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_TG3.cx"
TG3    <- process_cxfile(cxFile, "TG3", genome_metadata, genome_counts_filt, order_colors, col_name = "treatment")
```

```{r cor_TG3_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
TG3_pos_plot <- plot_graph(TG3$graph_pos, TG3$clusters_pos, "microbetag_analysis/figs/flashweave_scores/TG3_pos.svg")
TG3_neg_plot <- plot_graph(TG3$graph_neg, TG3$clusters_neg, "microbetag_analysis/figs/flashweave_scores/TG3_neg.svg")
```

### TG4

```{r cor_TG4, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_TG4.cx"
TG4    <- process_cxfile(cxFile, "TG4", genome_metadata, genome_counts_filt, order_colors, col_name = "treatment")
```


```{r cor_TG4_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
TG4_pos_plot <- plot_graph(TG4$graph_pos, TG4$clusters_pos, "microbetag_analysis/figs/flashweave_scores/TG4_pos.svg")
TG4_neg_plot <- plot_graph(TG4$graph_neg, TG4$clusters_neg, "microbetag_analysis/figs/flashweave_scores/TG4_neg.svg")
```

### TG5

```{r cor_TG5, comment="", message=FALSE, warning=FALSE}
cxFile <- "microbetag_analysis/microbetag_nets/cx/microbetag_net_TG5.cx"
TG5    <- process_cxfile(cxFile, "TG5", genome_metadata, genome_counts_filt, order_colors, col_name = "treatment")
```

```{r cor_TG5_clusters, message=FALSE, fig.height=8, fig.width=10, fig.fullwidth=TRUE}
TG5_pos_plot <- plot_graph(TG5$graph_pos, TG5$clusters_pos, "microbetag_analysis/figs/flashweave_scores/TG5_pos.svg")
TG5_neg_plot <- plot_graph(TG5$graph_neg, TG5$clusters_neg, "microbetag_analysis/figs/flashweave_scores/TG5_neg.svg")
```




```{r}
plot_graph <- function(graph, cluster, min_community_size = 2, vertex_color = NULL, vertex_size = NULL, edge_width = 1, mark_col = "#f4f4f4") {
  igraph::plot.igraph(
    graph,
    vertex.color = vertex_color %||% V(graph)$color,
    vertex.size = vertex_size %||% V(graph)$size,
    vertex.label = NA,
    vertex.frame.color = NA,
    edge.width = edge_width,
    mark.groups = communities(cluster) %>% keep(~ length(.x) >= min_community_size),
    mark.col = mark_col,
    mark.border = NA,
    layout = layout_with_fr
  )
}


# library(patchwork)
# 
# plots <- list(
#   TG1_neg_plot, TG1_pos_plot, TG2_neg_plot, TG2_pos_plot, TG3_neg_plot, TG3_pos_plot, TG4_neg_plot, TG4_pos_plot, TG5_neg_plot, TG5_pos_plot,
#   d7_neg_plot, d7_pos_plot, d14_neg_plot, d14_pos_plot, d21_neg_plot, d21_pos_plot, d28_neg_plot, d28_pos_plot, d35_neg_plot, d35_pos_plot
# )
# 
# grid_plot <- wrap_plots(plots, ncol = 5)
graphs <- list(
  TG1$graph_neg, TG1$graph_pos, TG2$graph_neg, TG2$graph_pos, TG3$graph_neg, TG3$graph_pos, TG4$graph_neg, TG4$graph_pos, TG5$graph_neg, TG5$graph_pos,
  day7$graph_neg, day7$graph_pos, day14$graph_neg, day14$graph_pos, day21$graph_neg, day21$graph_pos, day28$graph_neg, day28$graph_pos, day35$graph_neg, day35$graph_pos
)

names <- list(
  TG1$clusters_neg, TG1$clusters_pos, TG2$clusters_neg, TG2$clusters_pos, TG3$clusters_neg, TG3$clusters_pos, TG4$clusters_neg, TG4$clusters_pos, TG5$clusters_neg, TG5$clusters_pos,
  day7$clusters_neg, day7$clusters_pos, day14$clusters_neg, day14$clusters_pos, day21$clusters_neg, day21$clusters_pos, day28$clusters_neg, day28$clusters_pos, day35$clusters_neg, day35$clusters_pos
)

# Open a plotting device (PDF or PNG if saving)
pdf("microbetag_analysis/figs/flashweave_scores/4x5_graph_grid.pdf", width = 20, height = 20)

# Set up 5x5 plot layout
par(mfrow = c(4, 5), mar = c(0, 0, 0, 0))  # no margins

# Call plot_graph for each of the 25 cases
for (i in 1:20) {
  plot_graph(graphs[[i]], names[[i]])  # assuming lists of 25 graphs & clusters
}

dev.off()

```


## Combined

### Network properties

How the nodes in the graph tend to cluster together, indicating how tightly correlated groups of genomes are.

- **density**: the number of edges between nodes in a network. It is calculated by dividing the total number of edges in the network by the maximum number of possible edges.
- **modularity**: measures the strength of division of a network into modules. We use the `igraph::cluster_walktrap()`  to find densely connected subgraphs, also called communities in a graph via random walks. The idea is that short random walks tend to stay in the same community.
- **centrality:** measure the overall centralisation of a network and refers to the properties of the network structure as a whole
- 

```{r network_properties, comment="", message=FALSE, warning=FALSE}
bind_rows(day7$metrics_pos, 
          day14$metrics_pos, 
          day21$metrics_pos,
          day28$metrics_pos,
          day35$metrics_pos) %>% 
  tt()  # The tt() function renders a table in different formats 

bind_rows(day7$metrics_neg, 
          day14$metrics_neg, 
          day21$metrics_neg,
          day28$metrics_neg,
          day35$metrics_neg) %>% 
  tt()

bind_rows(TG1$metrics_pos,
          TG2$metrics_pos,
          TG3$metrics_pos,
          TG4$metrics_pos,
          TG5$metrics_pos)%>%
  tt()

bind_rows(TG1$metrics_neg,
          TG2$metrics_neg,
          TG3$metrics_neg,
          TG4$metrics_neg,
          TG5$metrics_neg,) %>%
  tt()
```

```{r transitivity_plots, comment="", message=FALSE, warning=FALSE}
plot_metrics <- function(metrics_list, group_var, group_levels) {
  bind_rows(metrics_list) %>%
    mutate(!!group_var := factor(.data[[group_var]], levels = group_levels)) %>%
    pivot_longer(!.data[[group_var]], names_to = "metric", values_to = "value") %>%
    filter(metric != "assortability") %>%
    ggplot(aes(x = .data[[group_var]], y = value)) +
      geom_col() +
      facet_wrap(. ~ metric, scales = "free", nrow = 1) +
      theme_classic() +
      theme(
        axis.text.x = element_text(angle = 45, hjust = 1),
        panel.background = element_rect(fill='transparent'),
        plot.background = element_rect(fill='transparent', color=NA),
        panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        legend.background = element_rect(fill='transparent'),
        legend.box.background = element_rect(fill='transparent')
      )
}

# Plot metrics: TREATMENTS POSITIVE
metrics_list <- list(TG1$metrics_pos, TG2$metrics_pos, TG3$metrics_pos, TG4$metrics_pos, TG5$metrics_pos)
treatment_levels <- c("TG1", "TG2", "TG3", "TG4", "TG5")
s <- plot_metrics(metrics_list, "treatment", treatment_levels)
ggsave("microbetag_analysis/figs/flashweave_scores/metrics_tp.svg", s, dpi = 300, device = "svg", bg = "transparent")


# Plot metrics: TREATMENTS NEGATIVE
metrics_list <- list(TG1$metrics_neg, TG2$metrics_neg, TG3$metrics_neg, TG4$metrics_neg, TG5$metrics_neg)
treatment_levels <- c("TG1", "TG2", "TG3", "TG4", "TG5")
r <- plot_metrics(metrics_list, "treatment", treatment_levels)
ggsave("microbetag_analysis/figs/flashweave_scores/metrics_tn.svg", r, dpi = 300, device = "svg", bg = "transparent")

# Plot metrics: DAYS NEGATIVE
metrics_list <- list(day7$metrics_neg, day14$metrics_neg, day21$metrics_neg, day28$metrics_neg, day35$metrics_neg)
treatment_levels <- c("day7", "day14", "day21", "day28", "day35")
q <- plot_metrics(metrics_list, "day", treatment_levels)
ggsave("microbetag_analysis/figs/flashweave_scores/metrics_dn.svg", q, dpi = 300, device = "svg", bg = "transparent")

# Plot metrics: DAYS POSITIVE
metrics_list <- list(day7$metrics_pos, day14$metrics_pos, day21$metrics_pos, day28$metrics_pos, day35$metrics_pos)
treatment_levels <- c("day7", "day14", "day21", "day28", "day35")
p <- plot_metrics(metrics_list, "day", treatment_levels)
ggsave("microbetag_analysis/figs/flashweave_scores/metrics_dp.svg", p, dpi = 300, device = "svg", bg = "transparent")
```





### Positive correlations

#### Clusters per day

```{r community_clusters_day, comment="", message=FALSE, warning=FALSE}
day_clusters_positive <- c(
  day7$communities_pos, 
  day14$communities_pos,
  day21$communities_pos,
  day28$communities_pos,
  day35$communities_pos
  ) %>%
  purrr::map(~ combn(.x, 2, simplify = FALSE)) %>%
  unlist(recursive = FALSE) %>% 
  purrr::map(~ sort(.x))

# Calculate functional distances
day_clusters_positive_dist <- compute_cluster_distances(day_clusters_positive, genome_gifts)

# Print top pairs
day_clusters_positive_dist %>% 
  filter(count>2) %>% 
  tt()
```

#### Clusters per treatment

```{r community_clusters_treatment, comment="", message=FALSE, warning=FALSE}
treatment_clusters_positive <- c(
  TG1$communities_pos, 
  TG2$communities_pos, 
  TG3$communities_pos, 
  TG4$communities_pos, 
  TG5$communities_pos
  ) %>%
  purrr::map(~ combn(.x, 2, simplify = FALSE)) %>%
  unlist(recursive = FALSE) %>% 
  purrr::map(~ sort(.x))

# Calculate functional distances
treatment_clusters_positive_dist <- compute_cluster_distances(treatment_clusters_positive, genome_gifts)

# Print top pairs
treatment_clusters_positive_dist %>% 
  filter(count>2) %>% 
  tt()
```

#### Cluster combined

```{r community_clusters_all, comment="", message=FALSE, warning=FALSE}
all_clusters_positive <- c(
  day7$communities_pos, 
  day14$communities_pos,
  day21$communities_pos,
  day28$communities_pos,
  day35$communities_pos
  ) %>%
  purrr::map(~ combn(.x, 2, simplify = FALSE)) %>%
  unlist(recursive = FALSE) %>% 
  purrr::map(~ sort(.x))

# Calculate functional distances
all_clusters_positive_dist <- compute_cluster_distances(all_clusters_positive, genome_gifts)

# Print top pairs
all_clusters_positive_dist %>% 
  filter(count>1) %>% 
  tt()
```

```{r community_clusters_all_plot, comment="", message=FALSE, warning=FALSE}
day_clusters_positive_filtered <- day_clusters_positive_dist %>% 
  filter(count>1) %>%
  rename(count_day=count) %>% 
  mutate(clustering_day=TRUE)

treatment_clusters_positive_filtered <- treatment_clusters_positive_dist %>% 
  filter(count>1) %>%
  rename(count_treatment=count) %>% 
  mutate(clustering_treatment=TRUE)

all_clusters_positive_filtered <- all_clusters_positive_dist %>% 
  filter(count > 1) %>%
  rename(count_all=count) %>% 
  mutate(clustering_all=TRUE)

associated_pairs_positive <- all_clusters_positive_filtered %>% 
  full_join(
    day_clusters_positive_filtered, 
    by=join_by("genomes"=="genomes","distance"=="distance")
  ) %>% 
  full_join(treatment_clusters_positive_filtered, by=join_by("genomes"=="genomes","distance"=="distance")) %>%
  rowwise() %>% 
  mutate(count=max(count_all,count_day,count_treatment, na.rm=TRUE)) %>%
  mutate(
    clustering = case_when(
      clustering_all == TRUE ~ "all",  # If clustering_all is TRUE
      clustering_day == TRUE & is.na(clustering_all) & is.na(clustering_treatment) ~ "day", 
      clustering_treatment == TRUE & is.na(clustering_all) & is.na(clustering_day) ~ "treatment"
    )
  ) %>%
  select(genomes,count,clustering,distance) %>%
  mutate(clustering=factor(clustering)) %>%

  # add coordinates
  separate(genomes, into = c("genome1", "genome2"), sep = " - ") %>%
  left_join(gift_pcoa_vectors %>% rownames_to_column(var = "genome"), by = c("genome1" = "genome")) %>%
  rename(x1 = Axis.1, y1 = Axis.2) %>%
  left_join(gift_pcoa_vectors %>% rownames_to_column(var = "genome"), by = c("genome2" = "genome")) %>%
  rename(x2 = Axis.1, y2 = Axis.2)

associated_pairs_positive %>% 
  select(genome1,genome2,count,clustering) %>% 
  left_join(genome_metadata %>% select(genome,order,species),by=join_by(genome1==genome)) %>% 
  left_join(genome_metadata %>% select(genome,order,species),by=join_by(genome2==genome)) %>% 
  tt()

```

### Negative correlations

#### Clusters per day

```{r community_clusters_day_negative, comment="", message=FALSE, warning=FALSE}
day_clusters_negative <- c(
  day7$communities_neg, 
  day14$communities_neg,
  day21$communities_neg,
  day28$communities_neg,
  day35$communities_neg
  ) %>%
  purrr::map(~ combn(.x, 2, simplify = FALSE)) %>%
  unlist(recursive = FALSE) %>% 
  purrr::map(~ sort(.x))

# Calculate functional distances
day_clusters_negative_dist <- compute_cluster_distances(day_clusters_negative, genome_gifts)

# Print top pairs
day_clusters_negative_dist %>% 
  filter(count>1) %>% 
  tt()
```

#### Clusters per treatment

```{r community_clusters_treatment_negative, comment="", message=FALSE, warning=FALSE}
treatment_clusters_negative <- c(
  TG1$communities_neg,
  TG2$communities_neg,
  TG3$communities_neg,
  TG4$communities_neg,
  TG5$communities_neg
  ) %>%
  purrr::map(~ combn(.x, 2, simplify = FALSE)) %>%
  unlist(recursive = FALSE) %>% 
  purrr::map(~ sort(.x))

treatment_clusters_negative_dist <- compute_cluster_distances(treatment_clusters_negative, genome_gifts)

# Print top pairs  --  all pairs actually appear only once
treatment_clusters_negative_dist %>% 
  filter(count>1) %>% 
  tt()
```

#### Cluster combined

```{r community_clusters_all_negative, comment="", message=FALSE, warning=FALSE}
all_clusters_negative <- c(
  day7$communities_neg,
  day14$communities_neg,
  day21$communities_neg,
  day28$communities_neg,
  day35$communities_neg,
  TG1$communities_neg,
  TG2$communities_neg,
  TG3$communities_neg,
  TG4$communities_neg,
  TG5$communities_neg
  ) %>%
  purrr::map(~ combn(.x, 2, simplify = FALSE)) %>%
  unlist(recursive = FALSE) %>% 
  purrr::map(~ sort(.x))

# Calculate functional distances
all_clusters_negative_dist <- compute_cluster_distances(all_clusters_negative, genome_gifts)

# Print top pairs
all_clusters_negative_dist %>% 
  filter(count>1) %>% 
  tt()
```

```{r community_clusters_all_plot2, comment="", message=FALSE, warning=FALSE}
day_clusters_negative_filtered <- day_clusters_negative_dist %>% 
  filter(count > 1) %>%
  rename(count_day=count) %>% 
  mutate(clustering_day=TRUE)

treatment_clusters_negative_filtered <- treatment_clusters_negative_dist %>% 
  filter(count > 1) %>%
  rename(count_treatment=count) %>% 
  mutate(clustering_treatment=TRUE)

all_clusters_negative_filtered <- all_clusters_negative_dist %>% 
  filter(count > 1) %>%
  rename(count_all=count) %>% 
  mutate(clustering_all=TRUE)

associated_pairs_negative <- all_clusters_negative_filtered %>% 
  full_join(day_clusters_negative_filtered, by=join_by("genomes"=="genomes","distance"=="distance")) %>% 
  full_join(treatment_clusters_negative_filtered, by=join_by("genomes"=="genomes","distance"=="distance")) %>%
  rowwise() %>% 
  mutate(count=max(count_all,count_day,count_treatment, na.rm=TRUE)) %>%
  mutate(clustering = case_when(
    clustering_all == TRUE ~ "all",  # If clustering_all is TRUE
    clustering_day == TRUE & is.na(clustering_all) & is.na(clustering_treatment) ~ "day", 
    clustering_treatment == TRUE & is.na(clustering_all) & is.na(clustering_day) ~ "treatment")) %>%
  select(genomes,count,clustering,distance) %>%
  mutate(clustering=factor(clustering)) %>%
  #add coordinates
  separate(genomes, into = c("genome1", "genome2"), sep = " - ") %>%
  left_join(gift_pcoa_vectors %>% rownames_to_column(var = "genome"), by = c("genome1" = "genome")) %>%
  rename(x1 = Axis.1, y1 = Axis.2) %>%
  left_join(gift_pcoa_vectors %>% rownames_to_column(var = "genome"), by = c("genome2" = "genome")) %>%
  rename(x2 = Axis.1, y2 = Axis.2)

associated_pairs_negative %>% 
  select(genome1, genome2, count, clustering) %>% 
  left_join(genome_metadata %>% select(genome,order,species),by=join_by(genome1==genome)) %>% 
  left_join(genome_metadata %>% select(genome,order,species),by=join_by(genome2==genome)) %>% 
  tt()
```

### Association vs. functional distance

```{r association_distance_pos_neg_test, comment="", message=FALSE, warning=FALSE}
weighted_sd <- function(x, w) {
  weighted_mean <- sum(w * x) / sum(w)
  variance <- sum(w * (x - weighted_mean)^2) / sum(w)
  sqrt(variance)
}

# Summary statistics
distance_stats <- bind_rows(
  associated_pairs_positive %>% mutate(type="positive"),
  associated_pairs_negative %>% mutate(type="negative")) %>% 
  left_join(genome_metadata %>% select(genome,genus),by=join_by(genome1==genome)) %>% 
  left_join(genome_metadata %>% select(genome,genus),by=join_by(genome2==genome)) %>%
  filter(genus.x!=genus.y) %>% #remove relationships with same genus due to crossmaping error probability
  group_by(type) %>% 
  summarise(
    weighted_mean = sum(distance * count) / sum(count),
    weighed_sd = weighted_sd(distance, count)
  )

distance_stats %>% 
  tt()

# Permutation test
weighted_mean <- function(x, w) {
  sum(x * w) / sum(w)
}

permutation_test <- function(data, n_permutations = 1000) {
  
  original_diff <- abs(
    weighted_mean(
      data$distance[data$type == "positive"], data$count[data$type == "positive"]
    ) - weighted_mean(
      data$distance[data$type == "negative"], data$count[data$type == "negative"]
    )
  )
  
  perm_diff <- replicate(n_permutations, {
    shuffled_type <- sample(data$type)
    abs(
      weighted_mean(
        data$distance[shuffled_type == "positive"], data$count[shuffled_type == "positive"]
      ) - weighted_mean(
        data$distance[shuffled_type == "negative"], data$count[shuffled_type == "negative"]
        )
    )
  })
  
  p_value <- mean(perm_diff >= original_diff)
  return(p_value)
}

permutation_test(bind_rows(associated_pairs_positive %>% mutate(type="positive"),
                                      associated_pairs_negative %>% mutate(type="negative")), 
                n_permutations = 10000)

```

In case you have not run previously run the [02_maga_catalogue.Rmd](./02_mag_catalogue.Rmd), you will need to 
get the `functional_distances` from there.

```{r association_distance_pos_background_test, comment="", message=FALSE, warning=FALSE}

# Calculate the weighted mean distance for the "positive" group
weighted_mean_positive <- sum(associated_pairs_positive$distance * associated_pairs_positive$count) / sum(associated_pairs_positive$count)

# Calculate the mean of all distances from the functional distance matrix
all_distances <- functional_distances[upper.tri(functional_distances, diag = FALSE)]
all_distances <- all_distances[!is.na(all_distances)]
sampled_distances <- sample(all_distances, size = 37, replace = FALSE)
mean_all_distances <- mean(sampled_distances)

# Number of permutations
n_permutations <- 10000

# Combine the positive distances and all distances into a single vector
combined_distances <- c(associated_pairs_positive$distance, sampled_distances)

# Get the observed difference in means
observed_diff <- weighted_mean_positive - mean_all_distances

# Permutation test
perm_diff <- replicate(n_permutations, {
  # Randomly shuffle the combined distances
  shuffled_distances <- sample(combined_distances)
  
  # Split into "positive" group (same size as original positive group) and "all" group
  shuffled_positive <- shuffled_distances[1:length(associated_pairs_positive$distance)]
  shuffled_all <- shuffled_distances[(length(associated_pairs_positive$distance) + 1):length(combined_distances)]
  
  # Calculate weighted mean for the shuffled positive group (use original counts)
  shuffled_weighted_mean_positive <- sum(shuffled_positive * associated_pairs_positive$count) / sum(associated_pairs_positive$count)
  
  # Calculate the mean for the shuffled all group
  shuffled_mean_all <- mean(shuffled_all)
  
  # Return the difference in means
  shuffled_weighted_mean_positive - shuffled_mean_all
})

# Calculate p-value (two-tailed test)
p_value <- mean(abs(perm_diff) >= abs(observed_diff))
p_value
```

```{r association_distance_neg_background_test, comment="", message=FALSE, warning=FALSE}
# Calculate the weighted mean distance for the "negative" group
weighted_mean_negative <- sum(associated_pairs_negative$distance * associated_pairs_negative$count) / sum(associated_pairs_negative$count)

# Calculate the mean of all distances from the functional distance matrix
all_distances <- functional_distances[upper.tri(functional_distances, diag = FALSE)]
all_distances <- all_distances[!is.na(all_distances)]
sampled_distances <- sample(all_distances, size = 37, replace = FALSE)
mean_all_distances <- mean(sampled_distances)

# Number of permutations
n_permutations <- 10000

# Combine the negative distances and all distances into a single vector
combined_distances <- c(associated_pairs_negative$distance, sampled_distances)

# Get the observed difference in means
observed_diff <- weighted_mean_negative - mean_all_distances

# Permutation test
perm_diff <- replicate(n_permutations, {
  # Randomly shuffle the combined distances
  shuffled_distances <- sample(combined_distances)
  
  # Split into "negative" group (same size as original negative group) and "all" group
  shuffled_negative <- shuffled_distances[1:length(associated_pairs_negative$distance)]
  shuffled_all <- shuffled_distances[(length(associated_pairs_negative$distance) + 1):length(combined_distances)]
  
  # Calculate weighted mean for the shuffled negative group (use original counts)
  shuffled_weighted_mean_negative <- sum(shuffled_negative * associated_pairs_negative$count) / sum(associated_pairs_negative$count)
  
  # Calculate the mean for the shuffled all group
  shuffled_mean_all <- mean(shuffled_all)
  
  # Return the difference in means
  shuffled_weighted_mean_negative - shuffled_mean_all
})

# Calculate p-value (two-tailed test)
p_value <- mean(abs(perm_diff) >= abs(observed_diff))
p_value


```

```{r association_distance_plot, comment="", message=FALSE, warning=FALSE}
tibble(dist=functional_distances[upper.tri(functional_distances, diag = FALSE)]) %>% 
  filter(!is.na(dist)) %>% 
  ggplot(aes(x = dist)) +
  geom_histogram(bins = 10, fill = "skyblue", color = "black", alpha = 0.7) +
  geom_vline(xintercept = pull(distance_stats[1,2]), linetype="dotted", color = "red", size=1.5) +
  geom_vline(xintercept = pull(distance_stats[2,2]), linetype="dotted", color = "green", size=1.5) +
  theme_minimal()
```

### Persistent associations

```{r community_clusters_all_combined_plot, comment="", message=FALSE, warning=FALSE, fig.height=6, fig.width=10, fig.fullwidth=TRUE}

connected_genomes <- unique(c(associated_pairs_negative$genome1, associated_pairs_negative$genome2,
                              associated_pairs_positive$genome1, associated_pairs_positive$genome2))

# reference_orders <- unique( genome_metadata %>% filter(genome %in%  connected_genomes) %>%  pull(order) )   # DEPR

# Add an alpha column: lower opacity for non-connected nodes
plot_data <- gift_pcoa_vectors %>%
  rownames_to_column(var="genome") %>%
  left_join(genome_metadata, by="genome") %>%
  mutate(alpha_value = ifelse(genome %in% connected_genomes, 1, 0.2)) %>%   # Lower opacity for non-connected
  mutate(auto=row_number())


Erysipelotrichales <- genome_metadata %>% filter(order == "o__Erysipelotrichales") %>% pull(genome)
UBA1381            <- genome_metadata %>% filter(order == "o__UBA1381") %>% pull(genome)
TANB77             <- genome_metadata %>% filter(order == "o__TANB77") %>% pull(genome)
Clostridiales      <- genome_metadata %>% filter(order == "o__Clostridiales") %>% pull(genome)
RF39               <- genome_metadata %>% filter(order == "o__RF39") %>% pull(genome)


Lachnospirales     <- genome_metadata %>% filter(order == "o__Lachnospirales") %>% pull(genome)
Oscillospirales    <- genome_metadata %>% filter(order == "o__Oscillospirales") %>% pull(genome)
Enterobacterales   <- genome_metadata %>% filter(order == "o__Enterobacterales") %>% pull(genome)
Christensenellales <-  genome_metadata %>% filter(order == "o__Christensenellales") %>% pull(genome)

genomes_of_taxa_of_interest <- c(
  # Oscillospirales
  # Lachnospirales
  # Enterobacterales, Christensenellales
  Erysipelotrichales, UBA1381, TANB77, Clostridiales, RF39
)


genomes_of_interest <- intersect(genomes_of_taxa_of_interest, connected_genomes)

# Make a df just for the needs of the text on the plot with the taxa name of interest
ps <- plot_data %>% filter(genome %in% genomes_of_interest)
ps <- ps %>% mutate(species = ifelse(species == "s__", genus, species))
key_value_text <- paste(ps$auto, ps$species, sep = ": ", collapse = "\n")

# Plot
q1 <- ggplot() +
    # Draws a straight line between points (x, y) and (xend, yend)


    geom_segment(
      data = associated_pairs_negative,
      aes(
        x = x1, y = y1, xend = x2, yend = y2, linetype=clustering
      ),
      color="#f5a2af",
      linewidth=1
    ) +
  
    geom_segment(
      data = associated_pairs_positive,
      aes(
        x = x1, y = y1, xend = x2, yend = y2, linetype=clustering
      ),
      color="#6AA84F",
      linewidth=1
    ) +
  
    scale_linetype_manual(values=c("solid", "dashed", "dotted")) +

    # Creates scatterplots
    geom_point(
      data = plot_data,
      aes(x = Axis.1, y = Axis.2, color = order, size = length, alpha = alpha_value),
      shape = 16) +

    geom_label(
      data = plot_data %>% filter(genome %in% genomes_of_interest),
      aes(x = Axis.1 , y = Axis.2 , label = auto),  # Adjust label position
      size = 3, 
      fill = "white", 
      color = "black",
      alpha = 0.31
    ) +  
    scale_alpha_identity() +  # Ensures the alpha values are used directly
    scale_color_manual(name = "Order", values = order_colors) +
    theme_minimal() +
    theme(legend.position = "right")

q1 <- q1 + guides(color = guide_legend(override.aes = list(size = 3)))  # Size of legend dots


q1 <- q1 + annotate(
        "text", x = 1, y = 6, 
        label = key_value_text,
        hjust = 0, vjust = 1, size = 3, color = "black"
    )

q1 <- q1 + 
  scale_x_continuous(expand = expansion(add = c(1.1, 0.3))) +  # Adjust x-axis expansion
  scale_y_continuous(expand = expansion(add = c(0.1, 0.1)))    # Adjust y-axis expansion


ggsave(
  "microbetag_analysis/figs/pcoa/pcoa.svg", q1, 
  dpi = 300, width = 9,  height = 9, units = "in", device = "svg", bg = "transparent"
) 




genomes_sets <- list(
  oscillospirales = Oscillospirales,
  lachnospirales = Lachnospirales,
  entero_christensenellales = c(Enterobacterales, Christensenellales),
  small = c(Erysipelotrichales, UBA1381, TANB77, Clostridiales, RF39)
)


for (set_name in names(genomes_sets)) {
  
  taxa <- genomes_sets[[set_name]]
  genomes_of_interest <- intersect(taxa, connected_genomes)
  
  ps <- plot_data %>% filter(genome %in% genomes_of_interest)
  ps <- ps %>% mutate(species = ifelse(species == "s__", genus, species))
  key_value_text <- paste(ps$auto, ps$species, sep = ": ", collapse = "\n")
  
  q1 <- ggplot() +
    geom_segment(
      data = associated_pairs_negative,
      aes(x = x1, y = y1, xend = x2, yend = y2, linetype=clustering),
      color="#f5a2af", linewidth=1
    ) +
    geom_segment(
      data = associated_pairs_positive,
      aes(x = x1, y = y1, xend = x2, yend = y2, linetype=clustering),
      color="#6AA84F", linewidth=1
    ) +
    scale_linetype_manual(values=c("solid", "dashed", "dotted")) +
    geom_point(
      data = plot_data,
      aes(x = Axis.1, y = Axis.2, color = order, size = length, alpha = alpha_value),
      shape = 16
    ) +
    geom_label(
      data = plot_data %>% filter(genome %in% genomes_of_interest),
      aes(x = Axis.1 , y = Axis.2 , label = auto),
      size = 3, fill = "white", color = "black", alpha = 0.31
    ) +
    scale_alpha_identity() +
    scale_color_manual(name = "Order", values = order_colors) +
    theme_minimal() +
    theme(legend.position = "right") +
    guides(color = guide_legend(override.aes = list(size = 3))) +
    annotate(
      "text", x = 1, y = 6, 
      label = key_value_text,
      hjust = 0, vjust = 1, size = 3, color = "black"
    ) +
    scale_x_continuous(expand = expansion(add = c(1.1, 0.3))) +
    scale_y_continuous(expand = expansion(add = c(0.1, 0.1)))
  
  ggsave(
    filename = paste0("microbetag_analysis/figs/pcoa/pcoa_", set_name, ".svg"),
    plot = q1,
    dpi = 300, width = 9, height = 9, units = "in",
    device = "svg", bg = "transparent"
  )
}
```



```{r relationship_complementarities1, comment="", message=FALSE, warning=FALSE, fig.height=14, fig.width=10, fig.fullwidth=TRUE}
selected_genomes <- c("GPB:bin_000009","GPB:bin_000021","GPB:bin_000208","D300452:bin_000016","D300444:bin_000010","GPB:bin_000140")

#Functional difference matrix
genome_gifts[selected_genomes,] %>%
    to.elements(., GIFT_db) %>%
    as.data.frame() %>%
    stats::dist(., method = "manhattan") / ncol(genome_gifts[selected_genomes, ])

genome_gifts[selected_genomes, ] %>% 
  to.elements(.,GIFT_db=GIFT_db) %>%
  as.data.frame() %>% 
  rownames_to_column(var="genome") %>% 
  pivot_longer(!genome, names_to = "trait", values_to = "gift") %>%
  mutate(functionid = substr(trait, 1, 3)) %>%
    mutate(trait = case_when(
      trait %in% GIFT_db$Code_element ~ GIFT_db$Element[match(trait, GIFT_db$Code_element)],
      TRUE ~ trait
    )) %>%
    filter(!functionid %in% c("D04","D08","B09","B10","S01","S02","S03")) %>% 
    mutate(functionid = case_when(
      functionid %in% GIFT_db$Code_function ~ GIFT_db$Function[match(functionid, GIFT_db$Code_function)],
      TRUE ~ functionid
    )) %>%
    mutate(genome=factor(genome,levels=selected_genomes)) %>% 
    mutate(trait=factor(trait,levels=unique(GIFT_db$Element))) %>%
    mutate(functionid=factor(functionid,levels=unique(GIFT_db$Function))) %>%
    ggplot(aes(x=genome,y=trait,fill=gift)) +
        geom_tile(colour="white", linewidth=0.2)+
        scale_fill_gradientn(colours=rev(c("#d53e4f", "#f46d43", "#fdae61", "#fee08b", "#e6f598", "#abdda4", "#ddf1da")))+
        facet_grid(functionid ~ ., scales="free",space="free") +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
              strip.text.y = element_text(angle = 0)) + 
        labs(y="Traits",x="Samples",fill="GIFT")
```

```{r relationship_complementarities2, comment="", message=FALSE, warning=FALSE, fig.height=14, fig.width=10, fig.fullwidth=TRUE}
selected_genomes <- c("GPB:bin_000208","D300452:bin_000016","TG5_28:bin_000003","D300509:bin_000001","GPB:bin_000123","GPB:bin_000124","GPB:bin_000194")

#Functional difference matrix
genome_gifts[selected_genomes,] %>%
    to.elements(., GIFT_db) %>%
    as.data.frame() %>%
    stats::dist(., method = "manhattan") / ncol(genome_gifts[selected_genomes, ])

genome_gifts[selected_genomes, ] %>% 
  to.elements(.,GIFT_db=GIFT_db) %>%
  as.data.frame() %>% 
  rownames_to_column(var="genome") %>% 
  pivot_longer(!genome, names_to = "trait", values_to = "gift") %>%
  mutate(functionid = substr(trait, 1, 3)) %>%
    mutate(trait = case_when(
      trait %in% GIFT_db$Code_element ~ GIFT_db$Element[match(trait, GIFT_db$Code_element)],
      TRUE ~ trait
    )) %>%
    filter(!functionid %in% c("D04","D08","B09","B10","S01","S02","S03")) %>% 
    mutate(functionid = case_when(
      functionid %in% GIFT_db$Code_function ~ GIFT_db$Function[match(functionid, GIFT_db$Code_function)],
      TRUE ~ functionid
    )) %>%
    mutate(genome=factor(genome,levels=selected_genomes)) %>% 
    mutate(trait=factor(trait,levels=unique(GIFT_db$Element))) %>%
    mutate(functionid=factor(functionid,levels=unique(GIFT_db$Function))) %>%
    ggplot(aes(x=genome,y=trait,fill=gift)) +
        geom_tile(colour="white", linewidth=0.2)+
        scale_fill_gradientn(colours=rev(c("#d53e4f", "#f46d43", "#fdae61", "#fee08b", "#e6f598", "#abdda4", "#ddf1da")))+
        facet_grid(functionid ~ ., scales="free",space="free") +
        theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1),
              strip.text.y = element_text(angle = 0)) + 
        labs(y="Traits",x="Samples",fill="GIFT")
```








```{r Seed scores}


# compl_scores <- read.table("microbetag_analysis/microbetag_nets/seed_scores.tsv", sep = "\t", header = TRUE)
compl_scores <- read.table("microbetag_analysis/mgg_prec/seeds_complementarity/phylomint_scores.tsv", sep = "\t", header = TRUE)

add_taxonomic_level <- function(df, tax_df, tax_level) {
  df <- df %>%
    left_join(tax_df %>% select(genome, all_of(tax_level)), by = c("A" = "genome")) %>%
    rename(!!paste0("A_", tax_level) := all_of(tax_level)) %>%
    left_join(tax_df %>% select(genome, all_of(tax_level)), by = c("B" = "genome")) %>%
    rename(!!paste0("B_", tax_level) := all_of(tax_level))
  return(df)
}


compl_scores_taxonomies <- add_taxonomic_level(compl_scores, genome_taxonomy, "phylum")
compl_scores_taxonomies <- add_taxonomic_level(compl_scores_taxonomies, genome_taxonomy, "order")
compl_scores_taxonomies <- add_taxonomic_level(compl_scores_taxonomies, genome_taxonomy, "family")
compl_scores_taxonomies <- add_taxonomic_level(compl_scores_taxonomies, genome_taxonomy, "genus")

```


```{r }
analyze_taxonomic_level <- function(df, tax_level) {
  tax_post_hoc <- df %>%
    group_by(!!sym(paste0("A_", tax_level)), !!sym(paste0("B_", tax_level))) %>%
    dplyr::tally() %>%  # Counts the number of entries per group
    left_join(
      df %>%
        # Group the dataset by the taxonomic level of interest for both A and B
        group_by(!!sym(paste0("A_", tax_level)), !!sym(paste0("B_", tax_level))) %>%

        # Apply a linear model (Complementarity ~ Competition) within each group
        dplyr::do(broom::glance(
          lm(Complementarity ~ Competition, data = .)
        )),
      by = c(paste0("A_", tax_level), paste0("B_", tax_level))
    ) %>%
    ungroup() %>%
    mutate(p.adjusted_bh = p.adjust(p.value, method = "BH")) %>%
    mutate(p.adjusted_bon = p.adjust(p.value, method = "bonferroni")) %>%
    mutate(p.adjusted_fdr = p.adjust(p.value, method = "fdr"))
  
  filtered_results <- tax_post_hoc %>% filter(adj.r.squared > 0 & nobs > 20)
  return(filtered_results)
}
phyla_post_hoc <- analyze_taxonomic_level(compl_scores_taxonomies, "phylum")
phyla_filtered_results <- phyla_post_hoc %>% filter(adj.r.squared > 0 & nobs > 20)

order_post_hoc <- analyze_taxonomic_level(compl_scores_taxonomies, "order")
order_filtered_results <- order_post_hoc %>% filter(adj.r.squared > 0 & nobs > 20)
```



```{r}
create_taxonomic_heatmap <- function(filtered_results, taxonomies, tax_level) {
  
  # Count occurrences at the chosen taxonomic level and create labels
  tax_counts <- taxonomies %>%
    count(!!sym(tax_level)) %>%
    mutate(label = paste(!!sym(tax_level), "(", n, ")", sep=""))

  # Merge the labels back to filtered_results
  filtered_results <- filtered_results %>%
    left_join(tax_counts %>% select(!!sym(paste0("A_", tax_level)) := !!sym(tax_level), 
                                    A_label = label), 
              by = paste0("A_", tax_level)) %>%
    left_join(tax_counts %>% select(!!sym(paste0("B_", tax_level)) := !!sym(tax_level), 
                                    B_label = label), 
              by = paste0("B_", tax_level))

  # Create the heatmap
  ggplot(
    filtered_results, 
    aes(
      x = A_label,
      y = B_label,
      fill = adj.r.squared
    )
  ) +
  geom_tile() +
  theme_minimal() +
  labs(
    x = paste("A", tax_level, sep = "_"),
    y = paste("B", tax_level, sep = "_"),
    fill = "adj.r.squared",
    title = paste("Predicting Complementarity as a function of Competition at", tax_level, "level")
  ) +
  scale_fill_gradient2(
      low = "white",       # Near 0
      high = "#3D85C6",  # Maximum value
      mid = "lightblue",       # Midpoint also white to ensure smooth transition
      midpoint = 0,        # Centered around 0
      na.value = "grey50",
      guide = "colourbar",
      aesthetics = "fill"
  ) +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  geom_text(
    aes(label = ifelse(p.adjusted_bh < 0.05, "*", "")), 
    color = "black"
  )
}

heatmap_phylum <- create_taxonomic_heatmap(phyla_filtered_results, genome_taxonomy, "phylum")
heatmap_order  <- create_taxonomic_heatmap(order_filtered_results, genome_taxonomy, "order")

# Set desired dimensions
svg("microbetag_analysis/figs/seed_stats/seed_scores_lm_order2.svg", width = 8, height = 6, bg="transparent")  

print(heatmap_order)  # Replace with your ggplot object
dev.off()  # Close the device to save the file
```
