==> 08-hmsc_analysis.Rmd <==
## Load data

```{r load_data_hmsc_analysis}
load("data/data.Rdata")
load("hmsc/fit_model1_250_10.Rdata")

# Select desired support threshold
support_threshold=0.9
negsupport_threshold=1-support_threshold
```

## Variance partitioning

```{r hmsc_variancepart, warning=FALSE, comments="", message=FALSE}

# Compute variance partitioning
varpart=computeVariancePartitioning(m)

varpart$vals %>%
   as.data.frame() %>%
   rownames_to_column(var="variable") %>%
   pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
   mutate(variable=factor(variable, levels=c("day","treatment","day:treatment","logseqdepth","Random: sample"))) %>%
   group_by(variable) %>%
   summarise(mean=mean(value)*100,sd=sd(value)*100) %>%
   tt()
```

```{r hmsc_varpart_plot, warning=FALSE, comments="", message=FALSE}
# Basal tree
varpart_tree <- genome_tree

#Varpart table
varpart_table <- varpart$vals %>%
   as.data.frame() %>%
   rownames_to_column(var="variable") %>%
   pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
   mutate(genome=factor(genome, levels=rev(varpart_tree$tip.label))) %>%
   mutate(variable=factor(variable, levels=rev(c("day","treatment","day:treatment","logseqdepth","Random: sample"))))

#Phylums
phylum_colors <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
  right_join(genome_metadata, by=join_by(phylum == phylum)) %>%
    filter(genome %in% varpart_tree$tip.label) %>%
    arrange(match(genome, varpart_tree$tip.label)) %>%
    mutate(phylum = factor(phylum, levels = unique(phylum))) %>%
    column_to_rownames(var = "genome") %>%
    select(phylum)


colors_alphabetic <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
  right_join(genome_metadata, by=join_by(phylum == phylum)) %>%
    filter(genome %in% varpart_tree$tip.label) %>%
    arrange(match(genome, varpart_tree$tip.label)) %>%
     select(phylum, colors) %>%
    unique() %>%
    arrange(phylum) %>%
    select(colors) %>%
    pull()

# Basal ggtree
varpart_tree <- varpart_tree %>%
        force.ultrametric(.,method="extend") %>%
        ggtree(., size = 0.3)

# Add phylum colors next to the tree tips
varpart_tree <- gheatmap(varpart_tree, phylum_colors, offset=-0.2, width=0.1, colnames=FALSE) +
   scale_fill_manual(values=colors_alphabetic)+
      labs(fill="Phylum")

#Reset fill scale to use a different colour profile in the heatmap
varpart_tree <- varpart_tree + new_scale_fill()

# Add variance stacked barplot
vertical_tree <-  varpart_tree +
       scale_fill_manual(values=c("#cccccc","#ed8a45","#be3e2b","#f6de6c"))+
        geom_fruit(
             data=varpart_table,
             geom=geom_bar,
             mapping = aes(x=value, y=genome, fill=variable, group=variable),
             pwidth = 2,
             offset = 0.05,
             width= 1,
             orientation="y",
             stat="identity")+
      labs(fill="Variable")

vertical_tree
```
## Model fit

```{r model_fit, warning=FALSE, comments="", message=FALSE}
MFCV <- evaluateModelFit(hM=m, predY=cv)

mean(MFCV$R2, na.rm = TRUE)

tibble(genome=m$spNames, r2 = MFCV[[2]]) %>% 
  pull(r2) %>% 
  hist()

predictive_genomes <- tibble(genome=m$spNames, r2 = MFCV[[2]]) %>% 
  filter(r2>=0.3)

# Abundance covered by predictive genomes
genome_counts_filt %>% 
mutate_if(is.numeric, ~ . / sum(.)) %>% 
right_join(predictive_genomes, by="genome") %>% 
  select(-c(genome,r2)) %>% 
  colSums() %>% 
  mean(., na.rm=TRUE)
```

```{r predictive_mags, warning=FALSE, comments="", message=FALSE}
var_pred_table <- tibble(mag=m$spNames,
       pred=MFCV$R2,
       var_pred=MFCV$R2 * varpart$vals[1,],
       support=getPostEstimate(hM=m, parName="Beta")$support %>% .[2,],
       estimate=getPostEstimate(hM=m, parName="Beta")$mean %>% .[2,]) %>%
  mutate(enrichment=ifelse(support>=support_threshold,"Feral","Neutral")) %>% 
  mutate(enrichment=ifelse(support<=negsupport_threshold,"Domestic",enrichment))

predictive_mags <- var_pred_table %>% 
  filter(var_pred>=0.005) %>% 
  pull(mag)
```

## Posterior estimates

```{r hmsc_postestimates, warning=FALSE, comments="", message=FALSE}
# Select desired support threshold
support=0.9
negsupport=1-support

# Basal tree
postestimates_tree <- genome_tree

# Posterior estimate table
post_beta <- getPostEstimate(hM=m, parName="Beta")$support %>%
    as.data.frame() %>%
    mutate(variable=m$covNames) %>%
    pivot_longer(!variable, names_to = "genome", values_to = "value") %>%
    mutate(genome=factor(genome, levels=rev(postestimates_tree$tip.label))) %>%
    mutate(value = case_when(
          value >= support ~ "Positive",
          value <= negsupport ~ "Negative",
          TRUE ~ "Neutral")) %>%
    mutate(value=factor(value, levels=c("Positive","Neutral","Negative"))) %>%
    pivot_wider(names_from = variable, values_from = value) %>%
    #select(genome,sp_vulgaris,area_semi,area_urban,sp_vulgarisxarea_semi,sp_vulgarisxarea_urban,season_spring,season_winter,sp_vulgarisxseason_spring,sp_vulgarisxseason_winter) %>%
    column_to_rownames(var="genome")

#Phylums
phylum_colors <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
  right_join(genome_metadata, by=join_by(phylum == phylum)) %>%
    filter(genome %in% postestimates_tree$tip.label) %>%
    arrange(match(genome, postestimates_tree$tip.label)) %>%
    mutate(phylum = factor(phylum, levels = unique(phylum))) %>%
    column_to_rownames(var = "genome") %>%
    select(phylum)


colors_alphabetic <- read_tsv("https://raw.githubusercontent.com/earthhologenome/EHI_taxonomy_colour/main/ehi_phylum_colors.tsv") %>%
  right_join(genome_metadata, by=join_by(phylum == phylum)) %>%
    filter(genome %in% postestimates_tree$tip.label) %>%
    arrange(match(genome, postestimates_tree$tip.label)) %>%
     select(phylum, colors) %>%
    unique() %>%
    arrange(phylum) %>%
    select(colors) %>%
    pull()

# Basal ggtree
postestimates_tree <- postestimates_tree %>%
        force.ultrametric(.,method="extend") %>%
        ggtree(., size = 0.3)

#Add phylum colors next to the tree tips
postestimates_tree <- gheatmap(postestimates_tree, phylum_colors, offset=-0.2, width=0.1, colnames=FALSE) +
      scale_fill_manual(values=colors_alphabetic)+
      labs(fill="Phylum")

#Reset fill scale to use a different colour profile in the heatmap
postestimates_tree <- postestimates_tree + new_scale_fill()

# Add posterior significant heatmap

postestimates_tree <- gheatmap(postestimates_tree, post_beta, offset=0, width=0.5, colnames=TRUE, colnames_position="top",colnames_angle=90, colnames_offset_y=1, hjust=0) +
        scale_fill_manual(values=c("#be3e2b","#f4f4f4","#b2b530"))+
        labs(fill="Trend")

postestimates_tree +
        vexpand(.25, 1) # expand top 
```
## Predict responses

```{r hmsc_temporal_predictions, warning=FALSE, comments="", message=FALSE}

# Select modelchain of interest
load("hmsc/fit_model1_250_10.Rdata")

gradient = c(7:35)
gradientlength = length(gradient)

#Treatment-specific gradient predictions
pred_TG1 <- constructGradient(m, 
                      focalVariable = "day", 
                      non.focalVariables = list(logseqdepth=list(1),treatment=list(3,"TG1")), 
                      ngrid=gradientlength) %>%
            predict(m, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(day=rep(gradient,1000)) %>%
            mutate(treatment=rep("TG1",gradientlength*1000)) %>%
            pivot_longer(-c(day,treatment), names_to = "genome", values_to = "value")

pred_TG2 <- constructGradient(m, 
                      focalVariable = "day", 
                      non.focalVariables = list(logseqdepth=list(1),treatment=list(3,"TG2")), 
                      ngrid=gradientlength) %>%
            predict(m, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(day=rep(gradient,1000)) %>%
            mutate(treatment=rep("TG2",gradientlength*1000)) %>%
            pivot_longer(-c(day,treatment), names_to = "genome", values_to = "value")

pred_TG3 <- constructGradient(m, 
                      focalVariable = "day", 
                      non.focalVariables = list(logseqdepth=list(1),treatment=list(3,"TG3")), 
                      ngrid=gradientlength) %>%
            predict(m, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(day=rep(gradient,1000)) %>%
            mutate(treatment=rep("TG3",gradientlength*1000)) %>%
            pivot_longer(-c(day,treatment), names_to = "genome", values_to = "value")

pred_TG4 <- constructGradient(m, 
                      focalVariable = "day", 
                      non.focalVariables = list(logseqdepth=list(1),treatment=list(3,"TG4")), 
                      ngrid=gradientlength) %>%
            predict(m, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(day=rep(gradient,1000)) %>%
            mutate(treatment=rep("TG4",gradientlength*1000)) %>%
            pivot_longer(-c(day,treatment), names_to = "genome", values_to = "value")

pred_TG5 <- constructGradient(m, 
                      focalVariable = "day", 
                      non.focalVariables = list(logseqdepth=list(1),treatment=list(3,"TG5")), 
                      ngrid=gradientlength) %>%
            predict(m, Gradient = ., expected = TRUE) %>%
            do.call(rbind,.) %>%
            as.data.frame() %>%
            mutate(day=rep(gradient,1000)) %>%
            mutate(treatment=rep("TG5",gradientlength*1000)) %>%
            pivot_longer(-c(day,treatment), names_to = "genome", values_to = "value")

pred_all <- rbind(pred_TG1,pred_TG2,pred_TG3,pred_TG4,pred_TG5)
```

### Function level

```{r hmsc_time_function_element, warning=FALSE, comments="", message=FALSE}
functions_table <- genome_gifts %>%
    to.elements(., GIFT_db) %>%
    to.functions(., GIFT_db) %>%
    as.data.frame()

community_functions <- pred_all %>%
  group_by(treatment, day, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    mutate(day=str_c(day,"_",treatment)) %>%
    dplyr::select(-c(row_id,treatment)) %>%
    column_to_rownames(var = "day") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(functions_table,.,GIFT_db) %>% 
    as.data.frame() %>%
    rownames_to_column(var="day") %>% 
    separate(day, into=c("day","treatment"), sep="_")
   })
```

### Biosynthesis

```{r hmsc_time_function_biosynthesis_plot, warning=FALSE, comments="", message=FALSE, fig.height=12, fig.width=10, fig.fullwidth=TRUE}
do.call(rbind, community_functions) %>% 
  pivot_longer(!c(day,treatment), names_to = "GIFT", values_to = "value") %>%
  mutate(day=as.numeric(day)) %>% 
  filter(GIFT %in% c("B02","B04","B07","B08")) %>% 
  ggplot(aes(x=day, y=value, group=day))+
    geom_boxplot(outlier.shape = NA) +
    facet_nested(GIFT ~ treatment, scales="free")
```

### Degradation

```{r hmsc_time_function_degradation_plot, warning=FALSE, comments="", message=FALSE, fig.height=12, fig.width=10, fig.fullwidth=TRUE}
do.call(rbind, community_functions) %>% 
  pivot_longer(!c(day,treatment), names_to = "GIFT", values_to = "value") %>%
  mutate(day=as.numeric(day)) %>% 
  filter(GIFT %in% c("D03","D05","D07","D09")) %>% 
  ggplot(aes(x=day, y=value, group=day))+
    geom_boxplot(outlier.shape = NA) +
    facet_nested(GIFT ~ treatment, scales="free")
```

### Element level

```{r hmsc_time_function_element2, warning=FALSE, comments="", message=FALSE}
elements_table <- genome_gifts %>%
    to.elements(., GIFT_db) %>%
    as.data.frame()

community_elements <- pred_all %>%
  group_by(treatment, day, genome) %>%
  mutate(row_id = row_number()) %>%
  pivot_wider(names_from = genome, values_from = value) %>%
  ungroup() %>%
  group_split(row_id) %>%
  as.list() %>%
  lapply(., FUN = function(x){x %>%
    mutate(day=str_c(day,"_",treatment)) %>%
    dplyr::select(-c(row_id,treatment)) %>%
    column_to_rownames(var = "day") %>%
    as.data.frame() %>%
    exp() %>%
    t() %>%
    tss() %>%
    to.community(elements_table,.,GIFT_db) %>% 
    as.data.frame() %>%
    rownames_to_column(var="day") %>% 
    separate(day, into=c("day","treatment"), sep="_")
   })

community_elements <- do.call(rbind, community_elements)
```
